Put purchase order handling in separate package
