+++ go.mod
@@ -1,0 +1,8 @@
+module po
+
+go 1.13
+
+require (
+	cloud.google.com/go/datastore v1.0.0
+	google.golang.org/api v0.15.0
+)
